fix issues from codecop
